ctype.h: make sure the input `signed char` values will get converted to 0..255
The MS clib expects those `isdigit/.../is<char_trait>` functions to
receive `int` in range -1..255, i.e. valid char should be 0..255, but
rest of the code is often dereferencing (signed) char pointer, making
the VS build sometimes getting into trouble on UTF8 source files, etc.

So this is complete review of the source, casting the values to 0..255
in the same style everywhere, making the code base signed/unsigned char
agnostic.

+++ sjasm/sjio.cpp
@@ -649,7 +649,7 @@ void ReadBufLine(bool Parse, bool SplitByColon) {
 			// copy the new character to new line
 			*rlppos = *rlpbuf++;
 			afterNonAlphaNum = afterNonAlphaNumNext;
-			afterNonAlphaNumNext = !isalnum(*rlppos);
+			afterNonAlphaNumNext = !isalnum((byte)*rlppos);
 			// Block comments logic first (anything serious may happen only "outside" of block comment
 			if ('*' == *rlppos && ReadBufData() && '/' == *rlpbuf) {
 				if (0 < blockComment) --blockComment;	// block comment ends here, -1 from nesting
